development/projects/GPT-Agents/chapter_2/requirements.txt
openai
python-dotenv
gpt4all